+++ db/migrate/20180910220523_create_states.rb
@@ -2,7 +2,7 @@ class CreateStates < ActiveRecord::Migration[5.2]
   def change
     create_table :states do |t|
       t.string :name
-      t.integer :max_hp
+      t.float :max_hp
       t.integer :water_need
       t.integer :soil_need
       t.string :img_url
